text_to_handwritting/requirements.txt
autopep8==2.0.1
Pillow==9.4.0
pycodestyle==2.10.0
